requirements.txt
numpy
scipy
termcolor
matplotlib
tqdm
progressbar
numpngw
tensorflow
tensorboardX
libtiff
pillow
tifffile
Cython
requests
itkwidgets
itk-io
sklearn
vispy
msgpack
argparse
jupyter
colored
tabulate
torchsummary
seaborn
qdarkstyle
dask
toolz
imagesize
SimpleITK
scikit-image
protobuf
tqdm
pyopengl
zarr
PyQtWebEngine
PyQt5-sip
numpy
pytest
pyqt5==5.15.7
pyqtgraph
pyqtwebengine
